refactor: Enhance lobby page structure with improved component separation
- Add isHost selector to GameSocketStore for centralized host status management
- Create new components:
  - StartButton: Extract game start logic and UI
  - InviteButton: Separate invitation functionality
- Refactor Setting component to use store's isHost selector
- Simplify LobbyPage to use new component structure
- Move START_BUTTON_STATUS to StartButton component
- Improve code organization and reusability
